fix: 토픽 페이지 구조 개선
