+++ apps/web/src/app/(with-side-bar)/topic/page.tsx
@@ -1,12 +1,5 @@
-import type { ContentTypeOptions } from "@/constants/content";
-import TopicBoardPage from "@/page/topic";
+import NewTopicPage from "@/page/topic";
 
-interface TopicBoardProps {
-  searchParams: Promise<{ id?: string; type?: ContentTypeOptions }>;
-}
-
-export default async function TopicBoard({ searchParams }: TopicBoardProps) {
-  const { id, type } = await searchParams;
-
-  return <TopicBoardPage id={id || ""} type={type || "ALL"} />;
+export default async function NewTopic() {
+  return <NewTopicPage />;
 }
