+++ apps/web/src/page/topic/[id]/sticker/index.tsx
@@ -1,0 +1,21 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+interface TopicStickerDetailPageProps {
+  id: string;
+}
+
+const BlockNote = dynamic(() => import("@/components/block-note"), { ssr: false });
+
+export default function TopicStickerDetailPage({ id }: TopicStickerDetailPageProps) {
+  return (
+    <main className="flex h-full flex-col gap-4">
+      <input
+        className="border-border w-full border-b pb-3 text-3xl font-bold outline-none"
+        placeholder="제목을 입력해주세요"
+      />
+      <BlockNote />
+    </main>
+  );
+}
